requirements-cupom-api.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.0.1
